Adding dto layer (#3)
* WIP - adding DTO layer. Initial trial with Brand entity to DTO

* WIP - Completed for the getById route. Need to update other endpoints so that entity is not directly exposed.

BrandDTO updated to relevant fields only.

Entity replaced with DTO

* WIP - Completed for the addBrand route.
Required opposite model mapper to convert DTO to entity.

* WIP - Completed for the updateBrand route.

Id added back into Brand entity to allow for mapping. Not sure if this is correct for production, but for the timebeing, it allows for updating.

* Implemented a get all brands route.

* class added for type checking.

* unused import removed.

* Updating fields to final/

* DTO layer started for Product entities. Implemented for the getById route.

* Added brand field into the ProductDTO and mapped the Brand name to it.

* ProductDTO implemented for the Add route, this now makes current Product endpoint more secure and de-coupled for the persistence layer of the app.

ProductDTO also updated with BrandDTO to pass brand id information and allow for easier mapping. This meant I could remove the custom mapping code from entityToDTO for brand related field.

* Get all products endpoint added using ProductDTO.

* ScraperResource Get All endpoint updated to use a DTO.

Some other considerations required as this entity and service is used in other object operations.